+++ src/source/RazorEngine.Core/Compilation/ReferenceResolver/IReferenceResolver.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Reflection;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RazorEngine.Compilation.ReferenceResolver
+{
+
+    public interface IReferenceResolver
+    {
+        IEnumerable<CompilerReference> GetReferences(TypeContext context, IEnumerable<CompilerReference> includeAssemblies = null);
+    }
+
+}