+++ src/source/RazorEngine.Core/Compilation/ReferenceResolver/UseCurrentAssembliesReferenceResolver.cs
@@ -3,28 +3,29 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Threading.Tasks;
 
-namespace RazorEngine.Compilation.Resolver
+namespace RazorEngine.Compilation.ReferenceResolver
 {
     /// <summary>
-    /// Resolves the assemblies by using all currently loaded assemblies. See <see cref="IAssemblyReferenceResolver"/>
+    /// Resolves the assemblies by using all currently loaded assemblies. See <see cref="IAssemblyResolver"/>
     /// </summary>
-    public class UseCurrentAssembliesReferenceResolver : IAssemblyReferenceResolver
+    public class UseCurrentAssembliesReferenceResolver : IReferenceResolver
     {
         /// <summary>
-        /// See <see cref="IAssemblyReferenceResolver.GetReferences"/>
+        /// See <see cref="IReferenceResolver.GetReferences"/>
         /// </summary>
         /// <param name="context"></param>
         /// <param name="includeAssemblies"></param>
         /// <returns></returns>
-        public IEnumerable<string> GetReferences(TypeContext context = null, IEnumerable<string> includeAssemblies = null)
+        public IEnumerable<CompilerReference> GetReferences(TypeContext context = null, IEnumerable<CompilerReference> includeAssemblies = null)
         {
             return CompilerServicesUtility
                    .GetLoadedAssemblies()
                    .Where(a => !a.IsDynamic && File.Exists(a.Location) && !a.Location.Contains(CompilerServiceBase.DynamicTemplateNamespace))
                    .GroupBy(a => a.GetName().Name).Select(grp => grp.First(y => y.GetName().Version == grp.Max(x => x.GetName().Version))) // only select distinct assemblies based on FullName to avoid loading duplicate assemblies
-                   .Select(a => a.Location)
-                   .Concat(includeAssemblies ?? Enumerable.Empty<string>());
+                   .Select(a => CompilerReference.From(a))
+                   .Concat(includeAssemblies ?? Enumerable.Empty<CompilerReference>());
         }
     }
 }